+++ Soulgram.Interests/Soulgram.Interests.Application/Commands/Movies/Handlers/CreateMovieCommandHandler.cs
@@ -2,7 +2,7 @@
 using Soulgram.Interests.Application.Commands.Genres;
 using Soulgram.Interests.Application.Interfaces;
 using Soulgram.Interests.Application.Models.Request;
-using Soulgram.Interests.Application.Queries;
+using Soulgram.Interests.Application.Queries.Genres;
 using Soulgram.Interests.Domain;
 
 namespace Soulgram.Interests.Application.Commands.Movies.Handlers;
@@ -33,7 +33,7 @@ public class CreateMovieCommandHandler : IRequestHandler<CreateMovieCommand>
         await CreateGenresIfNoExists(movie.GenresNames?.ToArray());
 
         await _repository.InsertOneAsync(movie, cancellationToken);
-        
+
         return Unit.Value;
     }
 
@@ -47,12 +47,9 @@ public class CreateMovieCommandHandler : IRequestHandler<CreateMovieCommand>
             .Select(g => g.Key)
             .ToArray();
 
-        if (notExistingGenres == null || !notExistingGenres.Any())
-        {
-            return;
-        }
-        
-        var bulkRequest = new CreateGenresBulkRequest()
+        if (notExistingGenres == null || !notExistingGenres.Any()) return;
+
+        var bulkRequest = new CreateGenresBulkRequest
         {
             GenreName = notExistingGenres
         };
