+++ Soulgram.Interests/Soulgram.Interests.Application/Queries/Interests/GetInterestQuery.cs
